+++ .secrets.baseline
@@ -76,10 +76,6 @@
       "path": "detect_secrets.filters.allowlist.is_line_allowlisted"
     },
     {
-      "path": "detect_secrets.filters.common.is_baseline_file",
-      "filename": ".secrets.baseline"
-    },
-    {
       "path": "detect_secrets.filters.common.is_ignored_due_to_verification_policies",
       "min_level": 2
     },
@@ -441,19 +437,10 @@
         "line_number": 627
       }
     ],
-    "lib/contracts/create3/utils_test.go": [
+    "lib/create3/create3_test.go": [
       {
         "type": "Hex High Entropy String",
-        "filename": "lib/contracts/create3/utils_test.go",
-        "hashed_secret": "87d1b1cb78352bf048e630bd5cbe69d35bdc313d",
-        "is_verified": false,
-        "line_number": 21
-      }
-    ],
-    "lib/contracts/utils_test.go": [
-      {
-        "type": "Hex High Entropy String",
-        "filename": "lib/contracts/utils_test.go",
+        "filename": "lib/create3/create3_test.go",
         "hashed_secret": "87d1b1cb78352bf048e630bd5cbe69d35bdc313d",
         "is_verified": false,
         "line_number": 22
@@ -813,5 +800,5 @@
       }
     ]
   },
-  "generated_at": "2024-03-22T13:22:59Z"
+  "generated_at": "2024-03-22T14:33:16Z"
 }
